requirements.txt
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.3.9
chardet==3.0.4
colorama==0.4.1
cx-Oracle==7.2.0
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-crispy-forms==1.7.2
django-heroku==0.3.1
gunicorn==19.9.0
idna==2.8
ipdb==0.12.2
ipython==7.7.0
ipython-genutils==0.2.0
jedi==0.14.1
lxml==4.3.3
MechanicalSoup==0.11.0
parso==0.5.1
pendulum==2.0.5
pickleshare==0.7.5
prompt-toolkit==2.0.9
psycopg2==2.8.3
Pygments==2.4.2
python-dateutil==2.8.0
pytz==2019.1
pytzdata==2019.2
requests==2.21.0
six==1.12.0
soupsieve==1.9
sqlparse==0.3.0
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
whitenoise==4.1.3
